requirements.txt
flask
line-bot-sdk
wheel
grpcio
google-cloud-firestore
flask_cors
line-bot-sdk
google-cloud-storage
pandas
gunicorn
google-cloud-logging
tensorflow
# tensorflow-cpu == 2.4.1
Pillow